personal_newsletter/LINE_Version/requirements.txt
line_bot_sdk==1.18.0
Flask==2.2.2
python-dotenv==0.20.0
